# [READY] Optimize strings comparison when sorting candidates