--- a/cpp/ycm/Candidate.cpp
+++ b/cpp/ycm/Candidate.cpp
@@ -18,44 +18,25 @@
 #include "Candidate.h"
 #include "Result.h"
 
-#include <algorithm>
-#include <locale>
-
-using std::all_of;
-
 namespace YouCompleteMe {
-
-static bool IsPrint( char c ) {
-  return std::isprint( c, std::locale::classic() );
-}
-
-
-static bool IsLower( char c ) {
-  return std::islower( c, std::locale::classic() );
-}
-
-
-bool IsPrintable( const std::string &text ) {
-  return all_of( text.cbegin(), text.cend(), IsPrint );
-}
-
 
 std::string GetWordBoundaryChars( const std::string &text ) {
   std::string result;
 
   for ( size_t i = 0; i < text.size(); ++i ) {
-    bool is_first_char_but_not_punctuation = i == 0 && !ispunct( text[ i ] );
+    bool is_first_char_but_not_punctuation = i == 0 &&
+                                             !IsPunctuation( text[ i ] );
     bool is_good_uppercase = i > 0 &&
                              IsUppercase( text[ i ] ) &&
                              !IsUppercase( text[ i - 1 ] );
     bool is_alpha_after_punctuation = i > 0 &&
-                                      ispunct( text[ i - 1 ] ) &&
-                                      isalpha( text[ i ] );
+                                      IsPunctuation( text[ i - 1 ] ) &&
+                                      IsAlpha( text[ i ] );
 
     if ( is_first_char_but_not_punctuation ||
          is_good_uppercase ||
          is_alpha_after_punctuation ) {
-      result.push_back( tolower( text[ i ] ) );
+      result.push_back( Lowercase( text[ i ] ) );
     }
   }
 
@@ -69,7 +50,7 @@ Bitset LetterBitsetFromString( const std::string &text ) {
   for ( char letter : text ) {
     int letter_index = IndexForLetter( letter );
 
-    if ( IsInAsciiRange( letter_index ) )
+    if ( IsAscii( letter_index ) )
       letter_bitset.set( letter_index );
   }
 
@@ -80,8 +61,9 @@ Bitset LetterBitsetFromString( const std::string &text ) {
 Candidate::Candidate( const std::string &text )
   :
   text_( text ),
+  case_swapped_text_( SwapCase( text ) ),
   word_boundary_chars_( GetWordBoundaryChars( text ) ),
-  text_is_lowercase_( all_of( text.cbegin(), text.cend(), IsLower ) ),
+  text_is_lowercase_( IsLowercase( text ) ),
   letters_present_( LetterBitsetFromString( text ) ),
   root_node_( new LetterNode( text ) ) {
 }
@@ -118,8 +100,8 @@ Result Candidate::QueryMatchResult( const std::string &query,
     index_sum += node->Index();
   }
 
-  return Result( true, &text_, text_is_lowercase_, index_sum,
-                 word_boundary_chars_, query );
+  return Result( true, &text_, &case_swapped_text_, text_is_lowercase_,
+                 index_sum, word_boundary_chars_, query );
 }
 
 } // namespace YouCompleteMe