--- a/cpp/ycm/IdentifierCompleter.cpp
+++ b/cpp/ycm/IdentifierCompleter.cpp
@@ -23,8 +23,6 @@
 #include "Utils.h"
 #include "ReleaseGil.h"
 
-#include <algorithm>
-
 namespace YouCompleteMe {