… cpp/ycm/tests/IsUppercase_test.cpp