After filtering the candidates, we sort them by following a number of rules and if none of them apply, we fall back to a lexicographic comparison where lower case letters are prioritized over upper case ones, which is the same as comparing them with inverted case. This is done with `std::lexicographical_compare` using the `ChangeCharCase` comparison function, which is called through `std::function`. Letter case is inverted in `ChangeCharCase` using `std::isupper`, `std::tolower`, and `std::toupper` with the locale parameter. This is inefficient:
 - we switch the case using the locale. In addition to be slow, this is incorrect because our strings are supposed to be encoded in UTF-8 (technically, we only support ASCII). For instance, we don't want to convert the byte `\xE9` (`é`) to `\xC9` (`É`) if the user locale is latin1 since these bytes are not valid UTF-8 sequences;
 - we unnecessarily call `std::function` on `ChangeCharCase`;
 - we compare the strings character by character while we could directly compare them with `<`.

We optimize this by storing the case-swapped text (i.e. the text where the lower case letters are converted to upper case ones and vice versa) as a property of the `Candidate` object and compare both candidates on this property with the `<` operator when falling back to lexicographic sorting. We also implement our own utility functions to test and convert character and string properties. This results in a massive performance improvement for the `CandidatesWithCommonPrefix` benchmark.

Before:
```
--------------------------------------------------------------------------------------------
Benchmark                                                     Time           CPU Iterations
--------------------------------------------------------------------------------------------
CandidatesWithCommonPrefix/1                                805 ns        800 ns     897430
CandidatesWithCommonPrefix/2                              10145 ns       9735 ns      64102
CandidatesWithCommonPrefix/4                              28211 ns      27743 ns      23617
CandidatesWithCommonPrefix/8                              65058 ns      65360 ns      11218
CandidatesWithCommonPrefix/16                            139190 ns     137666 ns       4986
CandidatesWithCommonPrefix/32                            289616 ns     277394 ns       2362
CandidatesWithCommonPrefix/64                            855752 ns     855086 ns        748
CandidatesWithCommonPrefix/128                          2235209 ns    2235683 ns        314
CandidatesWithCommonPrefix/256                          5533323 ns    5447654 ns        126
CandidatesWithCommonPrefix/512                         24811977 ns   24600158 ns         52
CandidatesWithCommonPrefix/1024                        29717991 ns   29900192 ns         24
CandidatesWithCommonPrefix/2048                        66555287 ns   65867089 ns          9
CandidatesWithCommonPrefix/4096                       147582755 ns  148200950 ns          4
CandidatesWithCommonPrefix/8192                       324108104 ns  327602100 ns          2
CandidatesWithCommonPrefix/16384                      704005910 ns  702004500 ns          1
CandidatesWithCommonPrefix/32768                     1694146814 ns 1700410900 ns          1
CandidatesWithCommonPrefix/65536                     3731511967 ns 3712823800 ns          1
IdentifierCompleterFixture_BigO                         3515.82 NlgN  3504.03 NlgN
IdentifierCompleterFixture_RMS                                  8 %          8 %             
```

After:
```
--------------------------------------------------------------------------------------------
Benchmark                                                     Time           CPU Iterations
--------------------------------------------------------------------------------------------
CandidatesWithCommonPrefix/1                                659 ns        654 ns    1121788
CandidatesWithCommonPrefix/2                                960 ns        960 ns     747858
CandidatesWithCommonPrefix/4                               1571 ns       1568 ns     407923
CandidatesWithCommonPrefix/8                               2753 ns       2719 ns     263950
CandidatesWithCommonPrefix/16                              5035 ns       5006 ns     112179
CandidatesWithCommonPrefix/32                             11132 ns      10951 ns      64102
CandidatesWithCommonPrefix/64                             21244 ns      21416 ns      32051
CandidatesWithCommonPrefix/128                            52992 ns      53435 ns      15473
CandidatesWithCommonPrefix/256                           124426 ns     125151 ns       7479
CandidatesWithCommonPrefix/512                           200050 ns     200287 ns       2804
CandidatesWithCommonPrefix/1024                          434089 ns     431772 ns       1662
CandidatesWithCommonPrefix/2048                          960146 ns     973485 ns        641
CandidatesWithCommonPrefix/4096                         2220949 ns    2186930 ns        321
CandidatesWithCommonPrefix/8192                         5132477 ns    5014318 ns        112
CandidatesWithCommonPrefix/16384                       11111614 ns   11212572 ns         64
CandidatesWithCommonPrefix/32768                       24127657 ns   24514443 ns         28
CandidatesWithCommonPrefix/65536                       51418317 ns   51480330 ns         10
IdentifierCompleterFixture_BigO                           49.01 NlgN    49.19 NlgN
IdentifierCompleterFixture_RMS                                1 %          2 %             
```
Algorithm is ~72 times faster in this particular scenario. Not bad.

Added a bunch of tests for the character/string utility functions.

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/valloric/ycmd/777)
<!-- Reviewable:end -->