@micbou 2017-06-11
I forgot to mention the increase in memory usage for adding the `_case_swapped_text` property. Let `n` be the number of candidates stored in memory. Since we only support text with at most 80 characters and [assuming an overhead of ~40 bytes from the `std::string` container](http://jovislab.com/blog/?p=76), we should expect an increase in memory usage of at most `120n` bytes. This gives us 120 MB for one million candidates. You probably can guess what happens to the ycmd process when storing as much candidates.

---

Reviewed 13 of 13 files at r1.
Review status: all files reviewed at latest revision, all discussions resolved.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/valloric/ycmd/777#-:-KmO23NkV07dQC5-QjsI:bks8s2j)*
<!-- Sent from Reviewable.io -->