@Valloric 2017-06-12
> Algorithm is ~72 times faster in this particular scenario. Not bad.

That's some good shit! :D

> We optimize this by storing the case-swapped text 

This is my only concern. As you pointed out in a separate comment, this will increase memory usage. We already have people complaining about ycmd RAM usage, but we've always traded off memory for speed and I'll take a 72x perf increase. RAM is cheap.

---

Review status: all files reviewed at latest revision, 3 unresolved discussions.

---

*[cpp/ycm/Utils.cpp, line 77 at r1](https://reviewable.io:443/reviews/valloric/ycmd/777#-KmSLkVhwgeTZVV22biJ:-KmSLkVhwgeTZVV22biK:b40ssy8) ([raw file](https://github.com/valloric/ycmd/blob/de0b54294deb9f1dfb004d405eb54c2d04a7e680/cpp/ycm/Utils.cpp#L77)):*
> ```cpp
> 
> bool IsPrintable( const std::string &text ) {
>   for ( char letter : text )
> ```

This should really have braces IMO.

---

*[cpp/ycm/Utils.cpp, line 100 at r1](https://reviewable.io:443/reviews/valloric/ycmd/777#-KmSZzL1CIbiPl3nU5Ge:-KmSZzL1CIbiPl3nU5Gf:brb8op1) ([raw file](https://github.com/valloric/ycmd/blob/de0b54294deb9f1dfb004d405eb54c2d04a7e680/cpp/ycm/Utils.cpp#L100)):*
> ```cpp
> // uppercase.
> bool IsLowercase( const std::string &text ) {
>   for ( char letter : text )
> ```

Braces

---

*[cpp/ycm/Utils.cpp, line 127 at r1](https://reviewable.io:443/reviews/valloric/ycmd/777#-KmS_0N88ssAUREiN2HT:-KmS_0N88ssAUREiN2HU:brb8op1) ([raw file](https://github.com/valloric/ycmd/blob/de0b54294deb9f1dfb004d405eb54c2d04a7e680/cpp/ycm/Utils.cpp#L127)):*
> ```cpp
> 
> bool HasUppercase( const std::string &text ) {
>   for ( char letter : text )
> ```

Braces

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/valloric/ycmd/777#-:-KmSLKBLrL0NJp-daZ21:bmg731t)*
<!-- Sent from Reviewable.io -->