@bstaletic 2017-06-13
What's up with the linux benchmark? It's not seeing nearly as big of a speed up.
```
Linux: IdentifierCompleterFixture/CandidatesWithCommonPrefix/65536 1313244016 - 1.3 s
OSX: IdentifierCompleterFixture/CandidatesWithCommonPrefix/65536  194837720 ns - ~0.2 s
Windows: IdentifierCompleterFixture/CandidatesWithCommonPrefix/65536  107713000 ns - ~0.1s
```

---

Reviewed 13 of 13 files at r1.
Review status: 12 of 13 files reviewed at latest revision, 4 unresolved discussions, some commit checks failed.

---

*[cpp/ycm/Utils.cpp, line 62 at r1](https://reviewable.io:443/reviews/valloric/ycmd/777#-KmTGaGh3p1-O8dxskuP:-KmTGaGh3p1-O8dxskuQ:b-vtndop) ([raw file](https://github.com/valloric/ycmd/blob/de0b54294deb9f1dfb004d405eb54c2d04a7e680/cpp/ycm/Utils.cpp#L62)):*
> ```cpp
> 
> bool IsAscii( char letter ) {
>   return !( letter & 0x80 );
> ```

This may need a comment. It does check that the eight bit in a `char` is 0, which is a correct way to do this, but I had to compile the line in my head (hex to binary), so may cause confusion if we ever decide to refactor these.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/valloric/ycmd/777#-:-KmTInAMobaJSgh4R0Tc:bl6sosv)*
<!-- Sent from Reviewable.io -->