@micbou 2017-06-13
Yes, the Linux benchmark is way slower than the OS X one on Travis but this was already the case before these changes:
```
Linux: IdentifierCompleterFixture/CandidatesWithCommonPrefix/65536 8763373879 ns - 8.8s
OS X: IdentifierCompleterFixture/CandidatesWithCommonPrefix/65536 2591216883 ns - 2.6s
```
The benchmark measurements are also a bit unreliable on Travis.

> This is my only concern. As you pointed out in a separate comment, this will increase memory usage. We already have people complaining about ycmd RAM usage, but we've always traded off memory for speed and I'll take a 72x perf increase. RAM is cheap.

I measured the memory usage from the benchmark using [Massif](http://valgrind.org/docs/manual/ms-manual.html) and obtained the following graphs:
 - [massif.out.master](https://gist.github.com/micbou/b13173a21392496d5a5c2e79e0b27383);
 - [massif.out.case-swapped-text](https://gist.github.com/micbou/fa58b3c56af93f4c904cd1252a255ba9).

There is a ~0.68% increase in memory usage. This is negligible.

---

Reviewed 1 of 1 files at r2, 1 of 1 files at r3, 2 of 2 files at r4.
Review status: all files reviewed at latest revision, 4 unresolved discussions.

---

*[cpp/ycm/Utils.cpp, line 62 at r1](https://reviewable.io:443/reviews/valloric/ycmd/777#-KmTGaGh3p1-O8dxskuP:-KmVer_YKxwu1SVbtcY4:bs3mhdg) ([raw file](https://github.com/valloric/ycmd/blob/de0b54294deb9f1dfb004d405eb54c2d04a7e680/cpp/ycm/Utils.cpp#L62)):*
<details><summary><i>Previously, bstaletic (Boris Staletic) wrote…</i></summary><blockquote>

This may need a comment. It does check that the eight bit in a `char` is 0, which is a correct way to do this, but I had to compile the line in my head (hex to binary), so may cause confusion if we ever decide to refactor these.
</blockquote></details>

Comment added.

---

*[cpp/ycm/Utils.cpp, line 77 at r1](https://reviewable.io:443/reviews/valloric/ycmd/777#-KmSLkVhwgeTZVV22biJ:-KmTgCem6o9vDHIIG7mU:b-896fix) ([raw file](https://github.com/valloric/ycmd/blob/de0b54294deb9f1dfb004d405eb54c2d04a7e680/cpp/ycm/Utils.cpp#L77)):*
<details><summary><i>Previously, Valloric (Val Markovic) wrote…</i></summary><blockquote>

This should really have braces IMO.
</blockquote></details>

Done.

---

*[cpp/ycm/Utils.cpp, line 100 at r1](https://reviewable.io:443/reviews/valloric/ycmd/777#-KmSZzL1CIbiPl3nU5Ge:-KmTgD8dpW267hnnlTPd:b-896fix) ([raw file](https://github.com/valloric/ycmd/blob/de0b54294deb9f1dfb004d405eb54c2d04a7e680/cpp/ycm/Utils.cpp#L100)):*
<details><summary><i>Previously, Valloric (Val Markovic) wrote…</i></summary><blockquote>

Braces
</blockquote></details>

Done.

---

*[cpp/ycm/Utils.cpp, line 127 at r1](https://reviewable.io:443/reviews/valloric/ycmd/777#-KmS_0N88ssAUREiN2HT:-KmTgDb8hJSBc9403bkU:b-896fix) ([raw file](https://github.com/valloric/ycmd/blob/de0b54294deb9f1dfb004d405eb54c2d04a7e680/cpp/ycm/Utils.cpp#L127)):*
<details><summary><i>Previously, Valloric (Val Markovic) wrote…</i></summary><blockquote>

Braces
</blockquote></details>

Done.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/valloric/ycmd/777#-:-KmVeUM5iNifxYfNpK6d:bf96xtw)*
<!-- Sent from Reviewable.io -->