@bstaletic 2017-06-13
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

---

Reviewed 2 of 2 files at r4.
Review status: all files reviewed at latest revision, 4 unresolved discussions.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/valloric/ycmd/777#-:-KmWB8asBxsjr_4fwAXq:bnfp4nl)*
<!-- Sent from Reviewable.io -->