@Valloric 2017-06-13
Wow, didn't expect memory growth to be so small!

<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/> and @zzbot r=bstaletic

---

Review status: all files reviewed at latest revision, 4 unresolved discussions.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/valloric/ycmd/777#-:-KmXiNxgu-ZccZYSop7z:bjmw4fp)*
<!-- Sent from Reviewable.io -->